+++ navbar-component.js
@@ -22,13 +22,7 @@ function Ctrl(
   $element, $rootScope, $scope, brAlertService, brNavbarService, config) {
   var self = this;
   self.brand = config.data.style.brand;
-  // default to showing navbar on route
-  self.route = $rootScope.route.current || {};
-  if(!('vars' in self.route)) {
-    self.route.vars = {navbar: true};
-  } else if(!('navbar' in self.route.vars)) {
-    self.route.vars.navbar = true;
-  }
+  self.route = $rootScope.route;
   self.siteTitle = config.data.siteTitle;
   self.service = brNavbarService;
   self.templates = [];